move message sender to public
