+++ message-receiver/include/Configuration/FileMessageCountingDecoratingReceiverConfiguration.h
