+++ message-receiver/include/Configuration/FileMessageCountingReceiverConfiguration.h
