+++ message-receiver/include/Configuration/FileMessageDecoratingReceiverConfiguration.h
