+++ message-receiver/include/Configuration/FileMessageReceiverConfiguration.h
