+++ message-receiver/include/Configuration/MessageCountingDecoratingReceiverConfiguration.h
