+++ message-receiver/include/Configuration/MessageCountingReceiverConfiguration.h
