+++ message-receiver/include/Configuration/MessageDecoratingReceiverConfiguration.h
