+++ message-receiver/include/MessageReceiver/MessageDecoratingFileReceiver.h
