+++ message-receiver/include/MessageReceiver/MessageReceiverFactory.h
@@ -25,9 +25,11 @@ class Notifier;
 
 class MessageReceiverFactory {
 public:
-	MessageReceiver* create(const MessageReceiverConfiguration & sender_configuration, CMSClientTestUnit & client_configuration, Notifier& parent)  const;
+	MessageReceiverFactory();
+	MessageReceiver* create(const MessageReceiverConfiguration & receiver_configuration, CMSClientTestUnit & client_configuration, Notifier& parent)  const;
 
 protected:
-	virtual MessageReceiver* create_receiver(const MessageReceiverConfiguration & sender_configuration, CMSClientTestUnit & client_configuration, Notifier& parent)  const;
-	virtual bool acceptedReceiverTypeType(const std::string &type) const;
+	virtual MessageReceiver* create_receiver(const MessageReceiverConfiguration & receiver_configuration, CMSClientTestUnit & client_configuration, Notifier& parent)  const;
+private:
+	const std::string mConsumerType;
 };
