+++ message-receiver/src/MessageReceiver/MessageFileReceiver.cpp
@@ -17,7 +17,7 @@
  * limitations under the License.
  */
 
-#include "MessageFileReceiver.h"
+#include <MessageReceiver/MessageFileReceiver.h>
 
 MessageFileReceiver::MessageFileReceiver(const FileMessageReceiverConfiguration & config, CMSClientTestUnit & client_params, Notifier & parent)
 	:MessageReceiver(config, client_params),
